Make flacoco not use test-runner
Implement all requires classes/interfaces ourselves. This allows us to
control the JVM startup and use a proper classpath.

test-runner had the problem that it did a lot of processing in the child
process (e.g. coverage parsing). This caused a lot of library conflicts
(e.g. the Lang defects4j project had an old version asm.jar included;
this meant that either test-runner or the Lang tests failed, depending
which jar won the classpath race...)