GeneticAlgorithm: Fix crash in crossover
When using LLM-mutations, it is possible that the block-structure of the
ASTs of two variants that are crossed over differs. What is a block in
one AST may be something else in the other AST. This leads to a crash,
if one is a leaf node.

Fix this by ensuring that we only add blocks (COMPOSIT_STATEMENT) to the
result of findMatchingModifiedBlocks.